requirements.txt
fastapi==0.111.1
uvicorn==0.24.0
langchain==0.0.189
transformers==4.33.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-multipart==0.0.6
numpy==1.26.4
